[refactor-ish] - gate dict autodownloading, lazy load to allow force_neural offline - save neural ckpt metadata, allow inference without dict (tags pos=UNK) - swap total_cost from int + scaled logprob to float + logprob (unsure how messy it makes api use, but i think this is more faithful probabilitistically) - some import tidying - add api/cli flags to toggle auto-dl & response type - span expansion should now correctly re-merge overlapping spans during fallback - add dev env skeleton, preliminary pytest testing: ckpt loading, fallback behavior, neural cost segmenting

+++ src/jp_tokenizer/types.py
@@ -28,5 +28,5 @@ class Token:
     feature: str
     start: int
     end: int
-    total_cost: int
+    total_cost: float
     source: str
